refactor(data): move entity-class mapping dicts to data/tables/
ENTITY_TYPE_TO_SA_CLASS, ENTITY_TYPE_TO_IRB_CLASS, and the derived
ENTITY_TYPES_BY_SA_CLASS encode regulatory taxonomy (CRR Art. 112,
147(3), 147(4)(b)) — they belong in data/tables/, not engine/. The
arch_check.py heuristic missed them because their RHS values are
ExposureClass.X.value (ast.Attribute), not raw string literals.

Relocates the three constants to a new
src/rwa_calc/data/tables/entity_class_mapping.py module and updates
the four engine import sites and four test import sites:

- engine/classifier.py: drops the three constants (-75 lines), imports
  ENTITY_TYPE_TO_SA_CLASS / ENTITY_TYPE_TO_IRB_CLASS from data/tables;
  the imports are at module level so the long-standing public path
  `from rwa_calc.engine.classifier import ENTITY_TYPE_TO_SA_CLASS`
  keeps working unchanged.
- engine/hierarchy.py, engine/sa/namespace.py, engine/irb/guarantee.py,
  engine/crm/guarantees.py: re-target imports at the source.
- tests/unit/test_high_risk_items.py, test_covered_bonds.py,
  test_sa_sl_classification.py, test_equity_routing.py: re-target
  imports at the source (the SL classification test splits its
  import to keep ExposureClassifier from engine.classifier).
- data/tables/__init__.py: re-exports the three names so they are
  also reachable via `from rwa_calc.data.tables import ...`.

This stacks on PR #309 (the classifier readability refactor) but is
independently correct — the dicts exist on master and the move is
purely mechanical.

Verified: arch_check, ruff, ty, public-import spot-check, and full
pytest suite (5778 passed, 2 skipped).
