+++ tests/unit/test_high_risk_items.py
@@ -40,11 +40,11 @@ from tests.fixtures.single_exposure import calculate_single_sa_exposure
 from rwa_calc.contracts.config import CalculationConfig
 from rwa_calc.data.tables.b31_risk_weights import B31_HIGH_RISK_RW
 from rwa_calc.data.tables.crr_risk_weights import HIGH_RISK_RW
-from rwa_calc.domain.enums import ExposureClass
-from rwa_calc.engine.classifier import (
+from rwa_calc.data.tables.entity_class_mapping import (
     ENTITY_TYPE_TO_IRB_CLASS,
     ENTITY_TYPE_TO_SA_CLASS,
 )
+from rwa_calc.domain.enums import ExposureClass
 from rwa_calc.engine.sa import SACalculator
 
 # =============================================================================
